requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
async-timeout==4.0.3
attributedict==0.3.0
attrs==21.4.0
auto_gptq==0.7.1
blessed==1.20.0
blessings==1.7
cachetools==5.3.3
categories==0.0.6
cbor2==5.6.4
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudscraper==1.2.71
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colorthief==0.2.1
colour-runner==0.1.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contourpy==1.3.1
coverage==7.5.2
cycler==0.12.1
DataProperty==1.0.1
datasets==3.0.1
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1733236420667/work
deepdiff==7.0.1
deepspeed==0.14.4
diffusion-core==0.0.68
dill==0.3.7
distlib==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.0
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1733327148154/work
exceptiongroup==1.2.1
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
filelock==3.14.0
filetype==1.2.0
flash-attn==2.7.3
fonttools==4.55.3
frozenlist==1.4.1
fsspec==2023.10.0
gekko==1.1.1
gitdb==4.0.12
GitPython==3.1.44
gpustat==1.1.1
grpcio==1.67.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.7
importlib_metadata==8.0.0
inspecta==0.1.3
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1732896932739/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.4
joblib==1.4.2
jsonlines==4.0.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
kiwisolver==1.4.8
lightning-utilities==0.11.3.post0
lxml==5.3.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mbstrdecoder==1.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.15
musicbrainzngs==0.7.1
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numexpr==2.10.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
-e git+https://github.com/OpenGVLab/OmniQuant.git@8eca0ce9ae222e344e5ce0d8897ed9e2b3404b6e#egg=omniquant
openai==1.30.3
ordered-set==4.1.0
packaging==24.0
pandas==2.2.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pathvalidate==3.2.0
peft==0.11.1
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.8.2
primefac==2.0.12
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1733302527033/work
protobuf==5.27.0
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pybind11==2.12.0
pycountry==23.12.11
pydantic==2.7.1
pydantic_core==2.18.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1733221634316/work
pyparsing==3.2.0
pyproject-api==1.6.1
pytablewriter==1.2.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytorch-lightning==2.3.1
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
regex==2024.5.15
requests==2.32.2
requests-toolbelt==1.0.0
rootpath==0.1.1
rouge==1.0.1
rouge_score==0.1.2
sacrebleu==1.5.0
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
screeninfo==0.8.1
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sqlitedict==2.1.0
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
stringcase==1.2.0
structlog==21.5.0
sympy==1.12
tabledata==1.3.3
tcolorpy==0.1.6
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
texttable==1.7.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
torch==2.3.0
torchmetrics==1.4.0.post0
torchvision==0.18.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827254365/work
tox==4.15.0
tqdm==4.66.4
tqdm-multiprocess==0.0.11
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.37.0
triton==2.3.0
typepy==1.3.2
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
wanda==1.0.1
wandb==0.19.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
Werkzeug==3.0.4
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
zstandard==0.22.0
